refactor : balanc => user 패키지로 변경(Aggregate Root)
