+++ src/main/java/kr/hhplus/be/server/api/user/domain/repository/BalanceRepository.java
@@ -1,6 +1,6 @@
-package kr.hhplus.be.server.api.balance.infrastructure;
+package kr.hhplus.be.server.api.user.domain.repository;
 
-import kr.hhplus.be.server.api.balance.domain.entity.Balance;
+import kr.hhplus.be.server.api.user.domain.entity.Balance;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
